backend/requirements.txt
fastapi
uvicorn
crewai
chromadb
pandas
langchain
langchain-community
google-generativeai
pydantic
python-dotenv
langchain-google-genai
ollama